Updated All  reimplemented Tests to match new structor
